% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Load}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{argument needed but automatically defined.}

\item{pkgname}{argument needed but automatically defined.}
}
\description{
This function should not be called by the user. It set initiate some hidden
variables used by this packages when the package is being loaded. The
function also modifies the prompt to indicate that packtrack is running.
}
\seealso{
\code{\link[base:ns-hooks]{ns-hooks}}
}
