% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onUnload}
\alias{.onUnload}
\title{Unload}
\usage{
.onUnload(libpath)
}
\arguments{
\item{libpath}{argument needed but automatically defined.}
}
\description{
This function should not be called by the user. It reset original R
parameters when the package is being unloaded.
}
