% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_stop.R
\name{packtrack_start}
\alias{packtrack_start}
\title{Start the tracking of namespaces}
\usage{
packtrack_start(create_cache = TRUE)
}
\arguments{
\item{create_cache}{whether or not to create a new .packtrack_data file (default = \code{TRUE}).}
}
\description{
This function activates packtrack and is automatically called when the package is loaded.
}
