% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_stop.R
\name{packtrack_stop}
\alias{packtrack_stop}
\title{Stop the tracking of namespaces}
\usage{
packtrack_stop(delete_cache = TRUE)
}
\arguments{
\item{delete_cache}{whether or not to delete the existing .packtrack_data file (default = \code{TRUE}).}
}
\description{
This function desctivates packtrack and is automatically called when the package is unloaded.
}
