% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{pktk_update}
\alias{pktk_update}
\title{Update the dictionary}
\usage{
pktk_update(pkg)
}
\arguments{
\item{pkg}{the name of the package to add to the dictionary}
}
\description{
This function either adds a new entry to the dictionary if the package is not among
the keys, or refreshes the time stamp otherwise.
}
