% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_data.R
\name{pktk_view}
\alias{pktk_view}
\title{View the data}
\usage{
pktk_view(previously_loaded = non_used, non_used = FALSE)
}
\arguments{
\item{previously_loaded}{boolean indicating whether to add (TRUE) or not
(FALSE) information about packages that were already loaded before
that the tracking was activated. This includes for instance the packtrack package
itself. Default is to do the same as the argument selected for non_used.}

\item{non_used}{a boolean indicating whether to add (TRUE) or not
(FALSE, default) information about packages not recorded during tracking.}
}
\description{
This function displays information about the usage of the packages.
}
\examples{
 pktk_view()

}
