% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pktk.R
\name{print.pktk}
\alias{print.pktk}
\title{Printing method for the object of class pktk}
\usage{
\method{print}{pktk}(x, ...)
}
\arguments{
\item{x}{the object to be printed}

\item{...}{not used but needed to define a printing method}
}
\description{
Printing method for the object of class pktk
}
