% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_stop.R
\name{packtrack_start}
\alias{packtrack_start}
\title{Start the tracking of namespaces}
\usage{
packtrack_start(replace_cache = FALSE)
}
\arguments{
\item{replace_cache}{whether or not to replace the existing .packtrack_data file (default = \code{FALSE}).}
}
\description{
This function activates packtrack and is automatically called when the package is loaded.
}
